requirements.txt
backtesting.py
bokeh
jupyter
pandas
numpy
seaborn
scikit-learn
supabase-py
redis
# postgres
psycopg
# graphql
python-dotenv
pyrh
tdameritrade
td-ameritrade-python-api
Backtesting==0.3.3
bokeh==2.4.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
Pygments==2.10.0
PyInquirer==1.0.3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
shellingham==1.4.0
six==1.16.0
termcolor==1.1.0
tornado==6.1
typer==0.3.2
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
wizardry==0.1.0
